Issue #741: Move ProcessBasedJobCreator to own submodule
